1) Some modifications to Kommandanter's recent land doctrine changes:
a) The Deep Battle nerf is re-done to be targeted more specifically at AT stacking. So, there is less HP for AT specifically, but infantry/art/mot HP is mostly back where it was. It is hoped that this makes AT bricks less attractive, in particular ones that stack excessive amounts of AT, but that infantry with reasonable amounts of ART and AT remain pretty good in DB. If people start stacking too much ART on DB instead, that may be lowered also back down a bit again later. The 5% HA that Kommandanter removed from DB AT remains gone.
b) The Grand Battleplan buff that Kommandanter made (10% extra breakthrough and 5% extra SA) is removed entirely, so GBP is back how it was before. This shouldn't necessarily make all that much difference; GBP should be still fairly OK for what people have recently liked about it, but hopefully it makes GBP not be too common and too cancerous again.
c) The slight mobile warfare buff Kommandanter put in remains the same.

2) AT nerfed a bit; Defence is lowered and also breakthrough is lowered of AT (the breakthrough part doesn't matter much because breakthrough is so low in the first place). The defense debuff is low enough that you should only notice it very much if you try to stack AT in particular; it won't make much difference for having a low-moderate amount of AT in your divisions.

3) Mountain terrain penalty is lowered from 50% attack to 45% attack. The idea here is to make mountains slightly more attackable so that we have less reliance on GBP as a way to attack mountains and less attractiveness of cancerous strategies that rely solely on trying to camp out in mountains forever on the defensive (fortress Libya too much, etc). Mountains should still be very good defensive terrain, but should be less unbreakable and hopefully can be taken to a reasonable degree if you do things like use mountaineers with assault guns etc, but not necessarily relying on GBP planning bonus for that (though it will help).