Added decisions for the Soviets that deploy up to a total of 418 unequipped divisions, 1 in each tile in the 4 front Soviet air zones (Baltic States, Belarus, Northern Ukraine, Southern Ukraine). These divisions deploy with 1 inf equip each, i.e. this gives Soviets a total of 418 free out-of-date inf equip while spawning the divisions. It is split into 4 decisions in case deploying too many at once might lag/crash the game (do your fellow players a favor and don't click on them all at once).
The decisions are only takeable between Jan 1940 and Sept 1940.

You will also want to change the template of the divisions once they deploy to whatever you are using.

Since it is against the rules to intentionally use unequipped combat divisions, obviously you have to build the equipment for any such divisions you deploy.

The purpose of this is simply to make deploying/spreading out divisions as the Soviets easier and quicker for the Soviets as a QOL improvement.