***Wakanda comes to the Pripjet Marshes of Belarus***
According to Galip's suggestion, as an experiment, a small amount of local supply from VPs is added to the interior of the Pripjet Marshes in Russia. This is not much, just adds up to a total of 12 local supply spread across 8 different tiles, but maybe it could keep a small # of infantry divisions alive for a bit, allowing for some swamp creature guerrilla war. If surrounded though, they should eventually run out of equipment if nothing else.