A lot of resources are moved from Sweden.
Axis minors get all of the Steel and Chromium (a lot of which is subsumed into Spain's steel focuses, other of which is moved into territory that Axis minors get/conquer). Some of the tungsten is moved to Austria and to Italy.

Incidentally, this also makes Asia matter a bit more because if Soviets run out of tungsten, they can't trade as much from Sweden any more.