Some nerfs/rebalancing to the talent trees for majors, particularly on the Allied/Soviet side. These seem to have been contributing to balance issues, in particular relating to Barbarossa, especially because:
1) The Soviet focuses were relatively good. Extra org offset a decent amount of the purge, while lower consumer goods boosted the already very large Allied economy. Apparently the AT dogs also gave HA to Soviet units other than, which I didn't realize.
2) Reduced production cost for fighters and reduced consumer goods/increased production efficiency cap meant that Soviets could get more air LL, and this also meant that the Allied air could be produced with less rubber, which meant that Japan could be ignored to a greater degree.

So, these focuses for majors in particular are nerfed to be more of minor/less important choices with smaller but real effects. It is fine for them to be more significant and have more buffs for minors, so generally with a few exceptions the minor trees remain basically as they were for i.e. countries like Australia/Canada/Hungary/Romania. The Raj is to some degree an exception because a very strong talent tree was contributing to the Tank Raj meta.

Also, a few other things are added. The best/most notable one is that Japan gets a choice to either Expand Trade with Australia or to Expand Trade with the Raj. "Expanding Trade" has the effect of building 3 new level 1 ports in that country... So Japan chooses which country should get 3 new ports to garrison... Which might make things interesting...