1) South Africa gets an Army Offense expert.
2) Thanks to some testing/data from IGB, Commonwealth and UK desert fox tank generals rebalanced so that they are all approximately equal in general modifiers (assuming they pick the best FM for Africa, which is generally an organizer logistics wizard). They each have slight differences, but overall it shouldn't be clear that there is any single country with the best tank general.

a) The UK general has the highest attack, but worst defense, logistics, and planning. This not coincidentally means that the UK general is especially well suited to Sealion defense, since in a Sealion, assuming UK has navy in the channel, it should have plenty of defense/breakthrough regardless, it is just a matter of how much attack can be stacked on UK tanks. There is also not much time to worry about planning bonus during a sealion, and UK should not have logistics problems in the UK, so a very high attack tank general will be best at kicking out a Sealion.

b) South Africa is fairly close to average in all traits, but a bit better (2nd best) in logistics, and 2nd worst in planning.

c) Canada is worst in attack, average in defense/brk, but best in both logistics and planning.

d) Raj is slightly above average in defense and planning and slightly below average in attack and logistics.

e) Australia is the best in defense, but 2nd worst in attack and slightly above average in logistics and planning.