Improvements in regards to naval balance - CV nerfs -  Slight doctrine alterations/irrelevant modifiers removed. (ie. 'mines_planting_by_fleets_factor)
Differences between the three doctrines have been exaggerated to be a bit more niche.  (Base Strike shouldn't be the doctrine choice every single game)

Base Strike:

-CV aircraft agility factor, as well as naval targeting reduced
-Capital ship modifiers reduced

Trade Interdiction:

-A bit more rewarding for submarines
-Penalty to enemy fleet from size ratio increased

Fleet in Being:

-Anti-air attack modifier increased

Generic CV doctrine techs:

-Naval targeting reduced by half

----------------------------------------------------

CV FIGHTERS - After days of testing I believe I've gotten them to a point worth fielding.  Take this with a grain of salt.  Of course there are several other variables to be accounted for in-game, but I would go as far as to say that a mixed carrier deck composition (i.e. 50/50 cv navs/fighters) is preferable to simply stacking 100% navs on your decks.

----------------------------------------------------

Carrier and Land-based aircraft should now sortie on the same tick***

----------------------------------------------------

The BASE hit chance for naval gunfire is no longer 100% (You read that correctly. It in fact was 100%.  Set now to 30%.  - Would like to lower to 10% - but changing gradually)

----------------------------------------------------

Damage reduction when unable to pierce armor set to 90% (99% --> 90%)

----------------------------------------------------