To adjust balance of Soviet winter offensives, I had been mostly adjusting the winter attack values (a small reduction to that was just reverted in the last commit).
However, it appears that this was probably the wrong variable, and the real problem is mostly that Axis org regain/recovery rate is too low in winter, especially when infra is damaged, because of recovery rate modifiers that stack additively.

So, the effect on recovery rate of snow/blizzard is lowered, and a define that alters recovery rate based on whether infrastructure is above/below 50% is lowered from changing it by 5% per level to 2.5% per level of infra. Basically what was apparently happening is Soviets would battleplan to damage infra, and when infra was damaged this made Axis troops recover org at a very low rate which stacked with the winter effects. That additive stacking of recovery rate leading to absurdly low or in some cases 0 recovery rate should be now less of a problem.

Hopefully we now have the right understanding/balancing of what was happening in winter.