IGB recently found that it was possible to "brute force" green air so that the other side's CAS was not flying, which has disrupted the air war and made it significantly more cancerous, with air controllers starting to switch air suddenly into a single air zone back and forth.
However, it appears that this problem is fixable by further defines/other changes. Specifically:

1) CAS and TAC "speed" is increased by 10x. This stat should not be thought of as having anything to do with actual speed, it is just a gameplay balancing value due to having to work around the defines we are stuck with, however much fun it may be to meme around about newly hypersonic CAS.
2) Air defense of all CAS/TAC reduced by a factor of 5. This seems to ensure that CAS/TAC still gets shot down by fighters at sufficiently high rates that you are not going to want to build only CAS with no fighters.
3) The DISRUPTION_DEFENCE_SPEED_FACTOR define is increased by a factor of 10. This makes planes that are very "fast" get disrupted less, so that CAS effectively seems to always joins battles (and both does direct damages and gives air support modifiers) as intended.
4) AIR_WING_MAX_STATS_SPEED is also increased by a factor of 10, because the effects of the DISRUPTION_DEFENCE_SPEED_FACTOR are scaled relative to the maximum allowed speed.
5) The ANTI_AIR_ATTACK_TO_AMOUNT define is reduced by a factor of 10. This ensures that CAS can still not get shot down by AA in troop templates, so you still can't do a no air build to shoot down CAS just by putting AA in your divisions.

This has not been tested yet with naval bomber disruption, so for the moment (if this is used in the Alizzaar server game today) there should be a rule that you are not allowed to use naval bombers, but also that the Allies are not allowed to raid the central Med and that Italy must put 5-10 factories on NAVs (which won't be used).

It is possible that NAVs might not get disrupted enough (or possibly get disrupted too much). So just don't mess with them for now until I or someone else gets a chance to test/check and ensure the balance is right on that.