Gibraltar is a new "terrain type." It is just like a mountain, except it is called "Gibraltar" instead.
When you are in the Gibraltar terrain type (which only exists on... you guessed it... Gibraltar! ...) there is only a single tactic that should ever fire in battles. This is the Gibraltar attack/defense tactic.

The Gibraltar attack/defense tactic raises combat width by 50%, so a typical 80 width battle becomes 120 width (hopefully 120 width is enough).

In addition, the Gibraltar tactic reduces damage done by both the attacker and the defender by 50%, which should mean that there are less crits and that battles go more slowly in Gibraltar. This should mean that both sides reinforce significantly better. Since invasions of Gibraltar will be slower, that also means that there is more time to try to convoy raid them.

Remember, not only the UK, but also the invaders can reinforce better with this change (and the invaders should also get critted less, since there is less damage). So UK may perhaps want to keep some navy around to raid any invasion of Gibraltar... which might even mean (gasp) that UK should build a dockyard or 2?

All the various mountain terrain buffs/debuffs should apply to Gibraltar, for everything from equipment/division types to generals.

Why 120 width? Mostly to try to avoid the situation where Axis then just invades with 80 width divisions or something. As it is, they could in theory invade with 60w marines, but if so they are more likely to get reinforce memed, and with reduced damage on both sides that hopefully shouldn't really make a difference regarding crits anyway. This may be revisited depending on how things go, and if it is a problem, units with width higher than 40 might end up being banned if need be.

Also, it is possible we might eventually consider doing something similar with a special terrain type/tactic for Stalingrad (Rattenkrieg Tactic!!!) and/or El Alamein, depending on how this goes.

Yes, making a new terrain type and a new special tactic just for that terrain type is a roundabout way to specially avoid reinforce memes in Gibraltar, but this appears to be the only way to do it AFAIK.