If you don't want scorched earth to be automated for Soviets, you can select in decisions to do it manually. This is not recommended, but is an option if you want it.
Note that if you do it manually, you will get the old decisions (which destroy a bit less infra in total than if it is automated - mainly with the automated scorched earth, a bit more infra gets destroyed if you push Soviets very far back and take Moscow/Stalingrad/Gorki/etc, though in that case you may have already lost or be close to losing). I am not going to bother to do the work to re-make the decisions with a bit more damaged infra just for those who want to do it manually, but if anyone else who wants to do it manually when they play Soviets feels like doing that work and sends me decision code that will damage the same infra as the automatic scorched earth, then I'll put that in.