An alternative way of treating higher draft laws, besides that old "service by requirement '42" system that got reverted, but also different from vanilla, so that people are not simply inclined to refuse to increase draft law and say GG in 1942 just because they run low on manpower.
Instead of getting industry penalties for high draft laws, you instead get some penalties (but also a few bonuses) for troops. Specifically, Service by Requirement / All Adults Serve / Scraping the Barrel give you:

-5 / - 10 / -15 max entrenchment
-.025 / -0.05 / -.075 CAS damage reduction
+.025 / +.05 / +.075 enemy air superiority bonus
- 1% / -2% / -3% attack and defense
+1.5% / +3% / +4.5% core attack and defense
+1.5% / +3% / +4.5% org recovery rate factor

a) The lower entrenchment is to represent that when you start scraping the barrel, you send your lower quality troops to defend static/low priority areas of the front (for example, low quality Osttruppen garrisoning the beaches of Normandy in 1944). This should make Axis more vulnerable in the late game to DDays and also to counterattacks on quiet parts of the front where they had been entrenched.

b) The CAS/air superiority stuff represents increasing late game vulnerability to enemy air superiority/CAS.

c) The lower attack/defense represents a general decline in the quality of troops (but is supposed to be small enough that it doesn't cause anyone to just GG). Also, this does NOT apply to tanks, special forces, or motorized divisions, but only to regular infantry divisions (there is no real drop off in the quality of the elite units, only in the reserves who have to start relying on grandfathers and young boys).

d) The core attack/defense bonus represents increasingly fanatical resistance in defending the homeland, but that won't help attacking Russia, it only helps in the Endsieg battles at home...

e) The increased recovery rate represents generally increased fanaticism.

The idea behind all this is that i.e. this will hurt German DDay garrisons in the late game in particular, but won't really hurt tank divisions that are attacking (although tanks will still be worse when entrenched defensively), and so Germany is less inclined to simply give up because they have to raise their draft law. Instead, Germany is encouraged to keep attacking, so maybe rather than just GGing they try for the Battle of Kursk/Battle of the Bulge and finally Steiner's counter-offensive. Also, having these changes to troop quality rather than industry reflects the historical reality that German production went *up* fairly substantially in 1944, when it was starting to really scrape the barrel.