1) Paratrooper cost is doubled.
2) New rule --- all paratroopers must be 40 width.

3) Germany starts with an extra 5k inf equip to make up for the extra cost of having paratroopers that are needed to paradrop places like Crete and possibly Benelux.

So, paratroopers are more expensive, and they need to be 40w. That means it is an investment to drop them, and if they die, you do lose something of some significance. So you should be careful and deliberate about when to make a paradrop.

As for people who are defending against paradrops, they should be aware that it is always possible the enemy may paradrop, and should take precautions accordingly. If the enemy does paradrop and you have guarded against it properly, take it as an invitation to have some free kills of enemy infantry that costs about 3x as much as regular infantry. Think of all the veterancy you can get by killing enemey paratroopers. See, it should be a GOOD thing to be paradropped, you should want that to happen, so that you can kill the paras and capture their equipment.