The air mission efficiency penalty for snow and blizzards is lowered to be the same as for heavy rain.
What was happening is that during winter, because of the weather, German planes in Russia would take mission efficiency penalties from weather. This tended to make German air and Soviet air suddenly be about equal in an unintended way (since in winter both sides get a mission efficiency debuff; Germany from snow/blizzard and Soviets from purge). Effectively, for purposes of air, winter was "the German purge."

As is, German air will still take some mission efficiency penalties during winter, just not as large, which should keep the air trades in Russia from not being equal during winter, and continuing to favor Germany, so that Soviets continue to be ground down for Fall Blau.

It is possible Soviet air might need to be buffed a bit at some point to make up for this, but leaving it off for now.