As it has been up to now, Germany gets better infrastructure in some occupied states, particularly in Benelux and France, than in Germany. This is problematic because it means that Germany is incentivized to build factories there rather than in Germany (despite 10% factories lost to occupation).
But even regardless of that, even if the infra were just as good in all of Germany as in for example Holland/Paris, Germany normally runs out of build slots in Germany itself fairly early, and ends up building a significant # of factories in occupied territory (particularly Benelux and France).

And this is problematic because it means that then Germany has i.e. 50 factories in Holland... And so then if the Allies DDay and take Holland, suddenly Germany has 50 less factories, and since USA/UK get direct control of captured territory, that also means the Allies get 50 more factories. That is quite a large industry swing, and it means that while trying to fight off a DDay Germany loses a bunch of efficiency on factories in an unrealistic way. That also makes it unrealistically good to try a DDay even if it is ultimately going to fail, because even if it fails, if you take that many factories, it is going to have a major impact.

Whereas realistically/historically, Germany would not be building its war industry in occupied territory (although it certainly did exploit foreign industry, in large part this was by bringing/forcing foreign workers/industry to work in Germany itself).

So...

1) A bunch of new states are added in Germany, to give more build slots to build in Germany. New states are added rather than simply making the states have more slots so that this should not really be a buff to Germany building infra, but rather just means that Germany should build in Germany without really changing the overall industry balance significantly.

2) Additionally, the infra is lowered in occupied France/Benelux. Partly it is lower at the start, but it can't be too low because there has to be enough supply during the Battle of France, so after France falls, during the Vichy events the infra in all of occupied Benelux/France should be (excepting any infra the AI builds, which might or might not be a problem - if it is, then probably more infra will be reduced/deleted in occupied territories in subsequent updates).