Deep Battle:
+0.5 HP for infantry
-1.5 HP for artillery

This is a bit of a nerf to ridiculous templates like Geometry's 10 artillery deep battle template (-13 HP or so for that), but should have a negligible effect on more reasonable templates (for example, +2 HP for the starting German/Soviet template with 3 ART/3 AT, or -0.5 HP for a 4 ART/4 AT template).

In and of itself this if a small change, but if templates that stack too much art (or AT) with DB continue to be too popular, minor adjustments can continue to be made and will eventually have an effect.

Just want to be careful not to overshoot to the point where people start making pure inf significantly more often (i.e. Gerardo has still used some pure 20w for delaying as Soviets, which while I don't think it is probably the best strategy, is not a tendency that is worth encouraging).