Africa stuff:
1) Egypt supply tweaked - it seems like the supply that Allies have gotten when Alpha built up Sinai seemed to result in pretty good battles for Egypt. However, it is not intended that UK should have to or be incentivized to spend a bunch of IC building up ports to improve supply above what it is supposed to be. Therefore:

a) Supply for the Allies is changed to be pretty close to the level it could be built up to by building up Sinai, but without requiring this buildup.
b) Furthermore, any additional buildup in either Sinai or Sudan should be impossible, because Sinai is moved into the Egypt supply zone and some additional ports are added in Suez, and a few other minor things.
c) When Allies lose Alexandria, their supply should not be screwed as much. This is due to extra ports in Sinai, and also extra small VPs and local supply in other places besides just Alexandria/Cairo.
d) Changes in supply on either side should be more gradual as Axis pushes forward into Egypt. This is due partly to some additional small VPs with small amounts of local supply, and partly to a slight increase in. This should mean that Axis should also have a bit more sufficient supply to attack (but definitely should have supply problems, and the balance should be Axis quality and supply issues vs Allied quantity and better supply, with Allied supply getting worse gradually as they are pushed back in Egypt).
e) Allied supply in Marsa Matruh should be a bit better, giving them a bit more supply advantage there.

Overall this should result in similar Africa gameplay to the last few games, which people have liked, but it should be smoother, with less sudden changes in supply for one side or another (though Alexandria and El Alamein will both still be very important).

Also hopefully more of the fight should occur in Marsa Matruh, and not just in Egypt itself; Allies should have a supply advantage to be able to fight there (and not really have more of a supply advantage in Egypt either), and since CAS joins regardless of red air, the air zone should not really be a reason to not fight there just because of the air zone (if that is still a problem, I will probably give UK upgraded range on 1936 fighters/CAS, and gives Germany some similar slightly upgraded range to balance it out in a subsequent update). But while Allies should be able to defend Egypt, we shouldn't want Allies to just roach in Egypt and never go into Marsa Matruh/Cyrenaica.

2) When Geometry's extra Soviet states were added, some airfields (including in Africa etc) apparently accidentally were moved. This is fixed (moved back).

3) Italian Libyan Infra focus gives level 10 infra in Cyrenaica; it is intended that Axis should not have to build anything up to attack (other than focus), and likewise Allies shouldn't have to build anything up to defend (other than focus).

4) Fixed minor but obscure Torch/Morocco issue that only @BaddoSpirito#5860 understands.