France starts without 1918 inf equip researched and can't research it (at least not until it tag switches). Why? To stop France from producing 1918 inf equip instead of 1936 inf equip so that Germany captures 1918 inf equip rather than 1936 inf equip when France caps.
Also, a general note --- if anyone tries any sort of "strategy" of deliberately not producing any equipment as France, just so that Germany doesn't capture anything, in that case this should be fixed after France caps by rehosting with an edited save in which a rough estimate of the equipment that France would have produced is added in.

Obvious exploits are... obvious... And remember that these things are against the rules, and not everything has to be literally spelled out in the rules.