Adding a thing from Expert AI mod that lets you view the stockpiles of other countries by sending a diplomatic request to view their stockpiles.
Currently this is not the best possible form for this (because you have to send a request and they have to accept it), so this might be improved later in a custom GUI where you can view your allies' stockpiles but not your enemies, but for now having it go through a diplomatic request seems like a significant improvement. So I am going ahead and including this initially as a first step as-is.

Thanks to Thrasymachus for letting me know about this possibility.