Bug Fixes (Free France):
While there are a couple of bugs that need debugging in France, this was a quick obvious one. When the new tag FRF was created, the history file was identical to FRA one, so a duplicate of all of the original generals/admirals was created. These all suck anyway, so deleting them is just decluttering, as no one in their right mind would use them.