-Altering Soviet and German Training Time modifiers to Minimum Training Level modifiers.
1.)  Soviet Union's -70% Training Time modifiers --> -70% Minimum Training Level

2.)  Germany's -50% Training Time modifier from Total War --> -50% Minimum Training Level

3.)  General Staff spirit now -15% Training Time down from -20%

This has been coming for a while, but for anyone who is unaware - The 'Training Time' modifier, whilst reducing the amount of time required for a unit to become 'deployable' in the training queue, also increases the rate in which a division gains experience, (i.e. Green, to Trained, to Regular, etc.)  This is a problem, as you can imagine.  German Units with these modifiers as an example will easily reach veterancy, and afterward also retain their veterancy against Allied divisions (who do not have this modifier) despite high losses - creating an untenable, and unreasonably difficult barrier to break through when lower level divisions engage in combat against veteran level ones.  Essentially a 'race' currently exists to see who can first gain veterancy.

This was not the original intention, as I do not believe this was known when these modifiers were given to these national Spirits. (Total War, War for Communism, etc.)  Instead, the modifier 'Minimum Training Level,' reduces the amount of time required to deploy units in queue, as originally intended.  (The same as the Training Time modifier, whilst not affecting a deployed units experience grind towards veterancy.)