Reverting for the moment at least Kommandanter's USA changes. As far as I can tell, there is a lack of a coherent rationale for why these changes are needed or desirable in terms of game balance. I asked for an explanation of the reasoning behind them, but still haven't gotten any answer. Insofar as I can understand the reasoning, some of it appears to be based on questionable premises, like the idea that the reason for doing suspend the persecution is to build early mils (as opposed to the efficiency cap having more of a benefit than the civ/infra construction advisor on your ultimate output, which should be the case at least for mils regardless).
The ways the PP etc were set up were based upon converting from vanilla by translating the congress costs into PP costs/bonuses when Congress was removed. While GDU does not fetishize staying close to vanilla, nevertheless if there is not a rationale/reason to 'stray' further from it, we shouldn't do so, since the timing of which focuses you can take when on USA should be similar to USA with the Congress mechanic, that makes it more playable/learnable for new players.

In practice, apparently the impact of this is a large buff to US industry (gives them about 40 more civs), and it is not clear why this is needed or desirable in terms of game balance. If USA is too weak, then if we want to buff them we should say "we are buffing USA because it is too weak" and make changes to do that, but it is not at all clear at least to me that USA is too weak. It seems in many ways quite strong in particular in the late game.

Komm's commit also apparently contains unrelated changes like changing the lengths of time for operations and some UK bonuses to ship production which are not mentioned in Komm's commit, much less given an explanation/rationale for why it is necessary to change them.

Exceptions that are not reverted:

a) Some localization fixes that he put in that were unrelated and not mentioned in the commit.
b) The changes related to weather effects on naval strikes (which don't really involve USA balance).