***Important Changes to how France Falls***
The goal is to improve the battle of France, allow the Allies to put up a good fight without the same fear of being trapped by unexpectedly early capitulation as all the territory around them flips to German control, and ensure that France has the proper incentive to make good equipment that is useful in doing as much damage to Germany (as opposed to make bad equipment so that Germany will capture useless equipment).

There is now an intermediate period of 10 days of "armistice negotiations" which start at around the time France would normally have fallen. Everyone gets notified that armistice negotiations have begun by a popup. As soon as this message is received, there is a 10 day countdown until France capitulates and the Vichy events start, so at that point Allies should get out of there (if 10 days is not enough time for Allies to reasonably get out, then the time will be adjusted/increased as needed).

"Armistice Negotiations" begin when France has more than 33% surrender progress and when it has lost a state (after going above 33% surrender progress). However, 33% surrender progress is basically the same time that France would previously have capitulated, because France's "Defended by the Maginot Line" national spirit is changed to make it not surrender (initially).

However, after the 10 days have passed, the "Defended by the Maginot Line" spirit changes so that what was formerly 33% surrender progress is now enough for France to instantly capitulate.

During the Armistice Negotiations period, fighting continues for a bit, but with less elan and less commitment to the fight. Why die for a war that is already over? So France gets a large penalty to attack during this period; however, it does not get a penalty to defense, so French troops can somewhat help with covering the retreat for the Allies as they all run away for the ports. Germany and the other Axis also get a smaller debuff to their attack values when attacking French troops, because likewise, they know the war is effectively over. No need to attack too ferociously and risk dying in a war that has already been won. So combat during this intermediate period should be comparatively low-intensity.

As you approach the end of the 10 day period, however (starting on about day 7), the French stockpile starts getting deleted and French troops start to disappear because they are strength deleted due to lack of equipment. This means that when France capitulates, Germany will capture essentially 0 equipment (Germany captures about 1 day's worth of production).

Instead, after another 10 days have passed, Germany will get a standard stockpile of captured French equipment added to its stockpile, which is the same no matter what sort of equipment France decides to produce. There is no possibility of trying to produce worthless equipment so that Germany captures that rather than good equipment (like tanks). So you may as well produce good equipment that is useful in killing as many Germans as possible and delaying France's capitulation as long as possible.

And at the end of the 10 day period, France capitulates.