The purge penalties for CAS mission efficiency and escort efficiency are switched.
What does this mean? It means that if Soviets don't escort their CAS with fighters, the CAS will get shot down quite a bit more (in some tests running the Saturday Game save with battleplan in Barbarossa, Soviets could lose i.e. 50 CAS per day in an air zone if there are no fighter escorts). However, if Soviets do escort with fighters, the fighters should be able to escort better and offset that more or less.

So basically, Soviets should actually need fighters more, and not only just CAS (although they will surely still need CAS).

If need be, this sort of thing can be extended to apply to non-Soviets, if it becomes more of an issue in general that people don't build fighters and only or mostly build CAS. But so far it seems like it has only/mostly been an issue for Soviets, because of the way the purge penalties were structured giving a higher penalty to fighters escorting CAS than to CAS itself.