CAS changes:
1) The base Ground Support bonus that troops get from CAS is lowered from .26 to .24; it seems like this should be safe since nobody really doubts that you need air/CAS at the moment, so no-air builds are not much of an immediate risk.

2) CAS/TAC air attack is reduced by about 3/4, while CAS/TAC air defense is reduced by about 1/2. The effect of this is that if you have German CAS against Soviet fighters, then Soviet fighters will trade favorably against the un-escorted CAS (even with purge). And more so if you have Allied fighters against German CAS (or German fighters against Allied CAS without fighter escorts).

Note that when I say 3/4 and 1/2, if you look at the files some of the #s will look a bit different as described, that is because of Komm's reversion. The 3/4 and 1/2 proprtions I reference are in comparison to what they just were, not to what they were a month or 2 ago.

3) However, the NAir.ESCORT_FACTOR define is increased from 2.6 to 7. This means that if you do have fighters escorting your CAS, this makes more of a difference. For example, if Germany has a bunch of fighters escorting its CAS, then Soviet fighters will no longer trade favorably against the German CAS (and German fighters will also massacre the Soviet fighters as normal during the purge).

Basically you need fighters to escort your CAS quite a bit more than before, because first of all if you don't have fighters the CAS should get shot down quite a bit more, and secondly if you *do* have fighters they will be more effective at escorting the CAS.

4) So that NAVs that have lots of fighter escorts should hopefully still be able to be disrupted, NAV bomber speed is lowered further (I didn't test this part at least so far, so the NAV speed might need some tweaking later).

In sum, people shouldn't overreact too much based on just a single game, in which it wasn't even entirely clear that Axis won only because of CAS. I don't think there is a basic conceptual issue with the way CAS works in GDU, it is just a question of tweaking the numbers to the right values, now that the meta has realized that CAS was comparatively much more useful than fighters (I would consider getting the meta to that point to be a positive achievement on the road to Valhalla). CAS can still always fly for both sides, and it isn't necessary to have 'hard forcing of green air.'

*In the 'hard forcing of green air' meta, CAS was interruptable (if you had enough fighters all concentrated in a single air zone), but it wasn't really escortable. Here, the ***opposite*** should more or less be the case - CAS should be escortable, but not interruptable.*

Note that this should also be a bit of a buff to Strat bombers and to a lesser extent TAC bombers, provided that they are escorted by a lot of fighters. For the moment at least I think that is probably ok, since TAC and strat were mostly useless and still are unlikely to be all that great for the moment. In fact, this may be a way to make strats useful which (if it can be done in a non-cancerous way) is something that GDU is really long-overdue for anyway. After some further practical game experience of how the escorting define is working in practice with CAS, we might look at extending things to strat/TAC a bit more.