Performance stuff.
For now just reducing the amount of stuff stored in combat logs, basically. In late game saves, records of combat seem to take up about 1/2 or more of the size of a save file, which is pretty huge.

Also, save files seem to contain numerous references to diplomacy/relations/opinion/etc for countries that do not actually exist, so possibly some performance gains could be had by deleting the extra AI countries/superfluous tags more thoroughly. However, that is not done for now because doing so seems to change country colors on the map in a weird way, but this is possible WIP for some future performance improvements..