1) CAS costs the same as fighters (CAS previously costed slightly less than fighters for some reason, although it also does cost a bit less than fighters in vanilla IIRC). This generally amounts to CAS being about 10% more expensive.
2) This define added ---

NDefines.NMilitary.LAND_AIR_COMBAT_MAX_PLANES_PER_ENEMY_WIDTH = 1 -- how many CAS/TAC can enter a combat depending on enemy width there

Previously GDU did not change this define from vanilla, in vanilla it is 3. When it is 3 in vanilla, that means that if a battle is 80 width, then 3x that (3 x 80 = 240), 240 CAS can join in each land battle.

So by lowering this to 1, that means that if a battle is 80 width, only 80 CAS at the most can join the battle. Basically this should make it less beneficial to spam huge amounts of CAS, since you will be capped out more easily in terms of how much CAS can join a battle by the combat width. You still should want to have enough CAS to join the battles, but it is a waste to have more than can fill the sum of the width of all battles going on in an air zone.

3) Ground support base modifier that troops get lowered again from 22.5% to 20%.